[hyperactor] factor `net` into modules
This is a straightforward refactor of `net` into `net::server` and `net::client`. More to come.

Differential Revision: [D86800094](https://our.internmc.facebook.com/intern/diff/D86800094/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86800094/)!

[ghstack-poisoned]
