+++ hyperactor/src/channel/net/framed.rs
@@ -18,6 +18,7 @@ use bytes::Buf;
 use bytes::BufMut;
 use bytes::Bytes;
 use bytes::BytesMut;
+use enum_as_inner::EnumAsInner;
 use futures::future::poll_fn;
 use tokio::io::AsyncRead;
 use tokio::io::AsyncReadExt;
@@ -310,6 +311,66 @@ impl<W: AsyncWrite + Unpin, B: Buf> FrameWrite<W, B> {
         match res {
             Ok(()) => Ok(writer),
             Err(e) => Err((writer, e)),
+        }
+    }
+}
+
+/// A state machine to manage cancellable writes. Notably, writes have
+/// values associated with them in order to maintain additional state
+/// for bookeeping purposes. WriteState ensures that an underlying
+/// `W`-typed writer is correctly owned by a [`FrameWrite`], or else
+/// is Idled.
+#[derive(EnumAsInner)]
+pub enum WriteState<W, F, T> {
+    /// No frame being written; the writer `W` is idle.
+    Idle(W),
+
+    /// Currently writing a frame, with associated T-typed bookeeping value.
+    /// The writer is owned by the [`FrameWrite`].
+    Writing(FrameWrite<W, F>, T),
+
+    /// Internal state to manage completions.
+    Broken,
+}
+
+impl<W, F: bytes::Buf, T> fmt::Debug for WriteState<W, F, T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            WriteState::Idle(_) => f.debug_tuple("Idle").finish(),
+            WriteState::Writing(frame_write, _) => {
+                f.debug_tuple("Writing").field(frame_write).finish()
+            }
+            WriteState::Broken => f.debug_tuple("Broken").finish(),
+        }
+    }
+}
+
+impl<W, F, T> WriteState<W, F, T> {
+    /// Current bookeeping value, if it exists.
+    pub fn value(&self) -> Option<&T> {
+        match self {
+            Self::Writing(_, v) => Some(v),
+            Self::Idle(_) | Self::Broken => None,
+        }
+    }
+}
+
+impl<W: AsyncWrite + Unpin, F: Buf, T> WriteState<W, F, T> {
+    /// Drive the send operation. (Cancellation safe.) This is a no-op
+    /// if the write state isn't sending anything; in this case, the operations
+    /// never returns, but may be canceled.
+    pub async fn send(&mut self) -> io::Result<T> {
+        match self {
+            Self::Idle(_) => futures::future::pending().await,
+            Self::Writing(fw, _value) => {
+                fw.send().await?;
+                let Ok((fw, value)) = std::mem::replace(self, Self::Broken).into_writing() else {
+                    panic!("illegal state");
+                };
+                *self = Self::Idle(fw.complete());
+                Ok(value)
+            }
+            Self::Broken => panic!("illegal state"),
         }
     }
 }
